requirements.txt
# Core dependencies
pydantic>=2.4.0
pydantic-settings>=2.0.0
typer>=0.9.0
rich>=13.3.5
python-dotenv>=1.0.0

# Database
SQLAlchemy>=2.0.0
alembic>=1.12.0
faiss-cpu>=1.7.4
nltk>=3.8.1

# API/Web
fastapi>=0.105.0
uvicorn>=0.23.0
streamlit>=1.31.0

# AI/ML
pydantic-ai>=0.7.0
langchain>=0.1.0
langchain-community>=0.0.10
openai>=1.1.0
anthropic>=0.6.0
sentence-transformers>=2.2.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1

# Utilities
tqdm>=4.66.0
loguru>=0.7.0
